+++ sigal/writer.py
@@ -24,7 +24,6 @@
 from __future__ import absolute_import
 
 import codecs
-import copy
 import jinja2
 import logging
 import os
@@ -34,9 +33,6 @@ from distutils.dir_util import copy_tree
 from jinja2 import Environment, FileSystemLoader, ChoiceLoader, PrefixLoader
 from jinja2.exceptions import TemplateNotFound
 
-import sigal.image
-import sigal.video
-from .settings import get_thumb, get_orig
 from .pkgmeta import __url__ as sigal_link
 
 THEMES_PATH = os.path.normpath(os.path.join(
@@ -46,17 +42,14 @@ THEMES_PATH = os.path.normpath(os.path.join(
 class Writer(object):
     """Generate html pages for each directory of images."""
 
-    def __init__(self, settings, output_dir, theme=None,
-                 template_file="index.html", output_file="index.html"):
-        self.settings = settings
-        self.output_dir = os.path.abspath(output_dir)
-        self.theme = theme or settings['theme']
-        self.template_file = template_file
-        self.output_file = output_file
-        self.logger = logging.getLogger(__name__)
+    template_file = 'index.html'
 
-        # optionally add index.html to the URLs
-        self.url_ext = self.output_file if settings['index_in_url'] else ''
+    def __init__(self, settings, theme=None, index_title=''):
+        self.settings = settings
+        self.output_dir = settings['destination']
+        self.theme = theme or settings['theme']
+        self.index_title = index_title
+        self.logger = logging.getLogger(__name__)
 
         # search the theme in sigal/theme if the given one does not exists
         if not os.path.exists(self.theme):
@@ -66,8 +59,8 @@ class Writer(object):
 
         self.logger.info("Theme  : %s", self.theme)
         theme_relpath = os.path.join(self.theme, 'templates')
-        default_loader = FileSystemLoader(
-            os.path.join(THEMES_PATH, 'default', 'templates'))
+        default_loader = FileSystemLoader(os.path.join(THEMES_PATH, 'default',
+                                                       'templates'))
 
         # setup jinja env
         env_options = {'trim_blocks': True}
@@ -92,121 +85,28 @@ class Writer(object):
             self.logger.error('The index.html template was not found.')
             sys.exit(1)
 
-        self.copy_assets()
-
-        self.ctx = {
-            'sigal_link': sigal_link,
-            'theme': {'name': os.path.basename(self.theme)},
-            'medias': [],
-            'albums': [],
-            'breadcrumb': ''
-        }
-
-    def copy_assets(self):
-        """Copy the theme files in the output dir."""
-
+        # Copy the theme files in the output dir
         self.theme_path = os.path.join(self.output_dir, 'static')
         copy_tree(os.path.join(self.theme, 'static'), self.theme_path)
 
-    def get_breadcrumb(self, paths, relpath):
-        """Paths to upper directories (with titles and links)."""
-
-        tmp_path = relpath
-        breadcrumb = [((self.url_ext or '.'), paths[tmp_path]['title'])]
-
-        while True:
-            tmp_path = os.path.normpath(os.path.join(tmp_path, '..'))
-            if tmp_path == '.':
-                break
-
-            url = os.path.relpath(tmp_path, relpath) + '/' + self.url_ext
-            breadcrumb.append((url, paths[tmp_path]['title']))
-
-        return reversed(breadcrumb)
-
-    def generate_context(self, paths, relpath):
+    def generate_context(self, album):
         """Generate the context dict for the given path."""
 
-        path = os.path.normpath(os.path.join(self.output_dir, relpath))
-        index_url = os.path.relpath(self.output_dir, path) + '/' + self.url_ext
-        self.logger.info("Output path : %s", path)
-
-        ctx = copy.deepcopy(self.ctx)
-        ctx.update({
+        self.logger.info("Output album : %s", album.dst_path)
+        return {
+            'album': album,
+            'index_title': self.index_title,
             'settings': self.settings,
-            'index_url': index_url,
-            'index_title': paths['.']['title']
-        })
-        ctx['theme']['url'] = os.path.relpath(self.theme_path, path)
+            'sigal_link': sigal_link,
+            'theme': {'name': os.path.basename(self.theme),
+                      'url': os.path.relpath(self.theme_path, album.dst_path)},
+        }
 
-        if relpath != '.':
-            ctx['breadcrumb'] = self.get_breadcrumb(paths, relpath)
-
-        if len(paths[relpath]['medias']) > 0 and self.settings['zip_gallery']:
-            ctx['zip_gallery'] = self.settings['zip_gallery']
-
-        for i in paths[relpath]['medias']:
-            media_ctx = {}
-            base, ext = os.path.splitext(i)
-            if ext in self.settings['img_ext_list']:
-                media_ctx['type'] = 'img'
-                media_ctx['file'] = i
-
-                file_path = os.path.join(path, i)
-                raw, simple = sigal.image.get_exif_tags(file_path)
-
-                if raw is not None:
-                    media_ctx['raw_exif'] = raw
-                if simple is not None:
-                    media_ctx['exif'] = simple
-            else:
-                media_ctx['type'] = 'vid'
-                media_ctx['file'] = base + '.webm'
-            media_ctx['thumb'] = get_thumb(self.settings, i)
-            if self.settings['keep_orig']:
-                media_ctx['big'] = get_orig(self.settings, i)
-            ctx['medias'].append(media_ctx)
-
-        for d in paths[relpath]['subdir']:
-            dpath = os.path.normpath(os.path.join(relpath, d))
-            alb_thumb = paths[dpath]['thumbnail']
-            thumb_name = get_thumb(self.settings, alb_thumb)
-            thumb_path = os.path.join(self.output_dir, dpath, thumb_name)
-            self.logger.debug("Thumbnail path : %s", thumb_path)
-
-            # generate the thumbnail if it is missing (if
-            # settings['make_thumbs'] is False)
-            if not os.path.isfile(thumb_path):
-                source = os.path.join(self.output_dir, dpath, alb_thumb)
-                ext = os.path.splitext(source)[1]
-                self.logger.debug("Generating thumbnail for %s", source)
-
-                if ext in self.settings['img_ext_list']:
-                    generator = sigal.image.generate_thumbnail
-                elif ext in self.settings['vid_ext_list']:
-                    generator = sigal.video.generate_thumbnail
-                else:
-                    generator = None
-                    self.logger.error('Unsupported file type for %s', source)
-
-                if generator:
-                    generator(source, thumb_path, self.settings['thumb_size'],
-                              fit=self.settings['thumb_fit'])
-
-            ctx['albums'].append({
-                'url': d + '/' + self.url_ext,
-                'title': paths[dpath]['title'],
-                'thumb': os.path.join(d, thumb_name)
-            })
-
-        return ctx
-
-    def write(self, paths, relpath):
+    def write(self, album):
         """Generate the HTML page and save it."""
 
-        ctx = self.generate_context(paths, relpath)
-        page = self.template.render(paths[relpath], **ctx)
+        page = self.template.render(**self.generate_context(album))
+        output_file = os.path.join(album.dst_path, album.output_file)
 
-        output_file = os.path.join(self.output_dir, relpath, self.output_file)
         with codecs.open(output_file, 'w', 'utf-8') as f:
             f.write(page)